issue #32 rename files
